introduce request chain support
It can happen that one HTTP ist not enough to reach termination.
Termination means a page that is final. In most cases this is a 200
response that contains the document/representation you care about. But
it can also be something from the 400-family (e.g. a 404 Not Found or a
410 Gone) or even a 500. However something from the 300-family is non
terminal, because it redirects to a different url. As we might
care about that as well, we will do additional request.
Currently this means up to 10 (hardcoded) requests.